+++ cmd/main.go
@@ -8,12 +8,10 @@ import (
 	"time"
 
 	postgres "github.com/Shalqarov/forum/internal/repository/postgres"
-	repository "github.com/Shalqarov/forum/internal/repository/sqlite"
 	"github.com/Shalqarov/forum/internal/session"
 	"github.com/Shalqarov/forum/internal/usecase"
 	"github.com/Shalqarov/forum/web"
 	_ "github.com/lib/pq"
-	_ "github.com/mattn/go-sqlite3"
 )
 
 func main() {
@@ -24,34 +22,26 @@ func main() {
 	infoLog := log.New(os.Stdout, "INFO\t", log.Ldate|log.Ltime)
 	errorLog := log.New(os.Stderr, "ERROR\t", log.Ldate|log.Ltime|log.Lshortfile)
 
-	dbConn, err := repository.OpenDB(*dsn)
+	dbconn, err := postgres.OpenDB(*dsn)
 	if err != nil {
 		errorLog.Fatal(err)
 	}
 	defer func() {
-		err := dbConn.Close()
-		if err != nil {
-			log.Fatal(err)
-		}
-	}()
-	postgre, err := postgres.OpenDB(*dsn)
-	if err != nil {
-		errorLog.Fatal(err)
-	}
-	defer func() {
-		err := dbConn.Close()
+		err := dbconn.Close()
 		if err != nil {
 			log.Fatal(err)
 		}
 	}()
 
 	router := http.NewServeMux()
-	userRepo := postgres.NewPostgresUserRepo(postgre)
-	postRepo := postgres.NewPostgresPostRepo(postgre)
-	commRepo := repository.NewSqliteCommentRepo(dbConn)
+	userRepo := postgres.NewPostgresUserRepo(dbconn)
+	postRepo := postgres.NewPostgresPostRepo(dbconn)
+	commRepo := postgres.NewSqliteCommentRepo(dbconn)
+	voteRepo := postgres.NewPostgresVoteRepo(dbconn)
 	userUsecase := usecase.NewUserUsecase(userRepo)
 	postUsecase := usecase.NewPostUsecase(postRepo)
 	commUsecase := usecase.NewCommentUsecase(commRepo)
+	voteUsecase := usecase.NewVoteUsecase(voteRepo)
 
 	templateCache, err := web.NewTemplateCache("./ui/html/")
 	if err != nil {
@@ -62,6 +52,7 @@ func main() {
 		UserUsecase:    userUsecase,
 		PostUsecase:    postUsecase,
 		CommentUsecase: commUsecase,
+		VoteUsecase:    voteUsecase,
 		TemplateCache:  templateCache,
 		ErrorLog:       errorLog,
 	})
