--- internal/repository/sqlite/setup.sql
@@ -1,47 +1,0 @@
--- users TABLE --
-CREATE TABLE IF NOT EXISTS "user" (
-    "id" INTEGER NOT NULL UNIQUE PRIMARY KEY AUTOINCREMENT,
-    "username" TEXT NOT NULL,
-    "email" TEXT NOT NULL UNIQUE,
-    "password" TEXT NOT NULL,
-    "avatar" TEXT NOT NULL
-);
--- post TABLE --
-CREATE TABLE IF NOT EXISTS "post" (
-    "id" INTEGER NOT NULL UNIQUE PRIMARY KEY AUTOINCREMENT,
-    "user_id" INTEGER NOT NULL,
-    "title" TEXT NOT NULL,
-    "content" TEXT NOT NULL,
-    "category" TEXT NOT NULL,
-    "date" TEXT NOT NULL,
-    "image" TEXT,
-    FOREIGN KEY (user_id) REFERENCES user(id)
-);
--- comment TABLE --
-CREATE TABLE IF NOT EXISTS "comment" (
-    "id" INTEGER NOT NULL UNIQUE PRIMARY KEY AUTOINCREMENT,
-    "user_id" INTEGER NOT NULL,
-    "post_id" INTEGER NOT NULL,
-    "content" TEXT NOT NULL,
-    "date" TEXT NOT NULL,
-    FOREIGN KEY (user_id) REFERENCES user(id),
-    FOREIGN KEY (post_id) REFERENCES post(id)
-);
--- post vote TABLE --
-CREATE TABLE IF NOT EXISTS "post_votes" (
-    "id" INTEGER NOT NULL UNIQUE PRIMARY KEY AUTOINCREMENT,
-    "user_id" INTEGER NOT NULL,
-    "post_id" INTEGER NOT NULL,
-    "vote" INTEGER NOT NULL,
-    FOREIGN KEY (user_id) REFERENCES user(id),
-    FOREIGN KEY (post_id) REFERENCES post(id)
-);
--- comment vote TABLE -- 
-CREATE TABLE IF NOT EXISTS "comment_votes"(
-    "id" INTEGER NOT NULL UNIQUE PRIMARY KEY AUTOINCREMENT,
-    "user_id" INTEGER NOT NULL,
-    "comment_id" INTEGER NOT NULL,
-    "vote" INTEGER NOT NULL,
-    FOREIGN KEY (user_id) REFERENCES user(id),
-    FOREIGN KEY (comment_id) REFERENCES comment(id)
-);
